refactor: extract createClientFromOpts helper to eliminate duplication
Move the repeated logger+auth+caCert+OctokitClient setup into a single
exported function createClientFromOpts() in init.ts. Both runOrgRepos()
in org-repos.ts and checkForMissingRepos() in main.ts now use this shared
helper, removing the duplicated wiring from each site.
